Make the batch inversion path float32-only (drop runtime dtype dispatch)
Replace the float32/float64 runtime dispatch (_invert_array2d_dispatch)
with a single deterministic float32 batch path, addressing PR review: the
kernel no longer chooses a code path from input dtype at run time.

- C++/SWIG: remove the double invert_array1d/invert_array2d batch entry
  points; the former *_f32 kernels take their names (float32 lives in the
  float* signature, not a name suffix). Single-pixel invert() stays double.
- Python: _require_float32 enforces float32 on imagery (raises on float64
  rather than a lossy float64->float32->double round-trip); the LUT is cast
  to float32 (it is sourced from the dual-use double interpolator). Coords,
  shade, solar angles, and results stay double.
- Tests: imagery fixtures are float32; add a float64-is-rejected test and a
  float32-determinism test; refresh the array1d doctest.

+++ tests/test_threading.py
@@ -25,9 +25,10 @@ expected_per_pixel = np.array(
 
 
 def _make_chunk(ny=8, nx=8):
+    # Imagery crosses the io->inversion boundary as float32 (contract dtype).
     n_bands = spectrum_target.size
-    targets = np.broadcast_to(spectrum_target, (ny, nx, n_bands)).copy()
-    backgrounds = np.broadcast_to(spectrum_background, (ny, nx, n_bands)).copy()
+    targets = np.broadcast_to(spectrum_target.astype(np.float32), (ny, nx, n_bands)).copy()
+    backgrounds = np.broadcast_to(spectrum_background.astype(np.float32), (ny, nx, n_bands)).copy()
     angles = np.full((ny, nx), solar_angle)
     return targets, backgrounds, angles
